app/src/main/java/com/candy/android/candyapp/helper/UiHelper.java
package com.candy.android.candyapp.helper;

import android.content.Context;
import android.util.DisplayMetrics;
import android.view.WindowManager;

/**
 * @author Marcin
 */

public class UiHelper {
    /**
     * This method converts dp unit to equivalent pixels, depending on device density.
     *
     * @param dp A value in dp (density independent pixels) unit. Which we need to convert into pixels
     * @param context Context to get resources and device specific display metrics
     * @return A int value to represent px equivalent to dp depending on device density
     */
    public static int convertDpToPixel(int dp, Context context){
        int px = (int) (dp * (context.getResources().getDisplayMetrics().densityDpi / 160f));
        return px;
    }

    /**
     * Return screen height
     * @param context Context to get window service.
     * @return A int value to represent screen height in pixels.
     */
    public static int getScreenHeight(Context context) {
        WindowManager manager = (WindowManager) context.getSystemService(Context.WINDOW_SERVICE);
        DisplayMetrics metrics = new DisplayMetrics();
        manager.getDefaultDisplay().getMetrics(metrics);
        return metrics.heightPixels;
    }
}
